% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packrn-package.R
\docType{package}
\name{packrn-package}
\alias{packrn}
\alias{packrn-package}
\title{packrn: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/NiloofarNL/packrn}
}
\author{
\strong{Maintainer}: Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
\concept{Assignment}
\keyword{internal}
